notes/orchard.rs mod
